+++ include/threadschedule/detail/pool/thread_pool_backend_base.hpp
@@ -10,6 +10,7 @@
 #include "../callable/bind.hpp"
 #include "../callable/move_only_function.hpp"
 #include "callbacks.hpp"
+#include "deadline.hpp"
 #include "indefinite_wait.hpp"
 #include "polling_wait.hpp"
 #include "shutdown_policy_backend.hpp"
@@ -424,7 +425,7 @@ public:
   {
     if (is_current_worker())
       detail::throw_worker_deadlock();
-    auto const deadline = std::chrono::steady_clock::now() + timeout;
+    auto const deadline = shutdown_deadline_after(timeout);
     std::lock_guard<std::recursive_mutex> shutdown_lock(shutdown_mutex_);
 
     std::unique_lock<std::mutex> lock(queue_mutex_);