+++ include/threadschedule/detail/scheduled/facade.hpp
@@ -8,7 +8,7 @@
 #include "../../worker_registration.hpp"
 #include "../pool/shutdown.hpp"
 #include "../thread/control.hpp"
-#include "time.hpp"
+#include "../time.hpp"
 
 #include <atomic>
 #include <chrono>