+++ include/threadschedule/detail/time.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../../expected.hpp"
+#include "../expected.hpp"
 
 #include <chrono>
 #include <cmath>